+++ charts/sonarqube/CHANGELOG.md
@@ -5,6 +5,9 @@ All changes to this chart will be documented in this file.
 * Update Chart's version to 2025.6.0
 * Upgrade SonarQube Community Build to 25.10.0.114319
 * Fix the `Service` selectors
+* Upgrade SonarQube Community Build to 25.11.0.114957
+* Disable Postgresql when the JDBC overwrite is used
+* Make SonarQube Server available as an Container App on the Azure Marketplace
 
 ## [2025.5.0]
 * Upgrade SonarQube Server to to 2025.5.0
@@ -12,6 +15,7 @@ All changes to this chart will be documented in this file.
 * Upgrade SonarQube Community Build to 25.9.0.112764
 * Update the image and readinessProbe used by postgresql after they migrated to a legacy repository
 * Support Kubernetes v1.34.0
+* Make SonarQube Server available as an Container App on the Azure Marketplace
 * Disable Postgresql when the JDBC overwrite is used
 
 ## [2025.4.0]